+++ custom_components/wirenboard_discovery/const.py
@@ -12,6 +12,20 @@ DEFAULT_PORT = 1883
 DEFAULT_PREFIX = "/"
 DEFAULT_SHOW_SYSTEM_DEVICES = False
 
-PLATFORMS = ["binary_sensor", "button", "climate", "cover", "number", "sensor", "switch", "text"]
+PLATFORMS = [
+    "binary_sensor",
+    "button",
+    "climate",
+    "cover",
+    "fan",
+    "light",
+    "lock",
+    "number",
+    "sensor",
+    "siren",
+    "switch",
+    "text",
+    "valve",
+]
 
 DISCOVERY_SECONDS = 4